Revert "Adjusted scripts"
This reverts commit bf881eec3fca53298991e6906497d4ea27edcba6.

+++ Thyme 2.0/Assets/Scripts/Combat/Classes/ComboSystem/Enemy/EnemyInfo.cs
@@ -13,7 +13,7 @@ public class EnemyInfo : MonoBehaviour
     public float launchSpeed = 5f;
     public float juggleForce = 1.5f;
 
-    public float health;
+    public int health;
 
     public bool inAir;
     public bool gettingLaunched;
@@ -63,7 +63,7 @@ public class EnemyInfo : MonoBehaviour
         hit = false;
     }
 
-    public void AdjustHealth(float i)
+    public void AdjustHealth(int i)
     {
         health -= i;
